fix: keep pasted image names browser safe
